+++ Additionals/ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## Version 1.8.1
 
-* ...
+* Tenth of seconds shown optionally only
 
 ## Version 1.8.0